worked on responsiveness and scrollbar
