+++ src/App.js
@@ -5,10 +5,11 @@ import SignUp from "./pages/SignUp/SignUp";
 import Waiting from "./pages/Waiting/Waiting";
 import Result from "./pages/Result/Result";
 import QuizClient from "./pages/QuizOngoing/quizpage";
-import Lounge from "./components/Lounge";
+import Lounge from "./components/Lounge/Lounge";
+import Login from "./components/Login";
 
 function App() {
-    return <Lounge/>;
+    return <SignUp/>;
 }
 
 export default App;
