+++ src/components/Login.css
@@ -15,7 +15,7 @@ body {
   margin: -20px 0 50px;
 }
 
-h1 {
+.crAcc {
   font-weight: bold;
   margin: 0;
 }
