+++ src/components/Lounge/Lounge.css
@@ -2,22 +2,35 @@
     color: #475edf;
     text-align: center;
 } 
+.main{
+    margin: 0px;
+}
  .hello{
-    position: absolute;
-    top:30px;
+    font-size: 5rem;
+    /* position: absolute;
+    top:2rem;
     left: 29%;
     font-size: 5rem;
-    padding: 2rem;
+    padding: 2rem; */
+}
+.box{
+    position: absolute;
+    top:12%;
+    right: 32%;
+    /* margin: 0px; */
 }
 .waitingText{
-    position: absolute;
+    /* position: absolute;
     padding: 30px;
     top:25%;
-    right:35%
+    right:35% */
 }
 .msg{
     
-    color: black;
+    color:#475edf;
+}
+.msgTitle{
+    color: #475edf;
 }
 .srol{
     height: 100%;
@@ -25,17 +38,10 @@
     overflow-y: scroll;
     scrollbar-width: 10px;
 }
-.scroll::-webkit-scrollbar{
-    width: 10px;
-  }
-.scroll::-webkit-scrollbar-thumb{
-    border-radius: 50px;
-    background-color: #475edf;
-    background: linear-gradient(#576ce0,rgb(113, 122, 218),#475edf );
-}
+
 .scroll::-webkit-scrollbar-track{
     backdrop-filter: blur(50px);
-  }
+}
     
 .msgTitle{
     /* padding:20px 10px 10px 10px; */
@@ -71,7 +77,7 @@
     right: 0px;
     top: 0px;
     margin: 0px;
-    padding: 0;
+    padding: 10px;
     backdrop-filter: blur(40px);
 }
 .textbox{
@@ -87,5 +93,32 @@
     border-radius: 30%;
 
 }
-
-
+.aninmateit{
+    position: absolute;
+    transform: translateX(180deg);
+    height: 100%;
+    width: 30%;
+    bottom: 0rem;
+    left: 0rem;
+}
+@media (max-width: 900px) {
+    
+    .box{
+        position: absolute;
+        top :2rem;
+        right: 1rem;
+        padding-left:35px ;
+        font-size: 0.75rem;
+    }
+    .timer{
+        position: static;
+        display:inline-flex;
+        right: 20px;
+        
+    }
+    .timerspn{
+        position: relative;
+        top: 230px;
+        left: 14%;        
+    }
+  }
